Add enrichment-rules to client and dynakube
