+++ config/crd/bases/dynatrace.com_dynakubes.yaml
@@ -4567,6 +4567,23 @@ spec:
                 description: KubeSystemUUID contains the UUID of the current Kubernetes
                   cluster
                 type: string
+              metadataEnrichment:
+                description: Observed state of Metadata-Enrichment
+                properties:
+                  rules:
+                    items:
+                      properties:
+                        enabled:
+                          type: boolean
+                        key:
+                          type: string
+                        mapping:
+                          type: string
+                        type:
+                          type: string
+                      type: object
+                    type: array
+                type: object
               oneAgent:
                 description: Observed state of OneAgent
                 properties:
