+++ pkg/clients/dynatrace/client.go
@@ -79,6 +79,8 @@ type Client interface {
 	// or an api error otherwise
 	GetSettingsForMonitoredEntities(ctx context.Context, monitoredEntities []MonitoredEntity, schemaId string) (GetSettingsResponse, error)
 
+	GetRulesSetting(ctx context.Context) (GetRulesSettingsResponse, error)
+
 	GetActiveGateAuthToken(ctx context.Context, dynakubeName string) (*ActiveGateAuthTokenInfo, error)
 
 	GetLatestOneAgentImage(ctx context.Context) (*LatestImageInfo, error)
