+++ pkg/clients/dynatrace/settings_kubernetes.go
@@ -31,10 +31,10 @@ type MonitoringSettings struct {
 }
 
 type postKubernetesSettingsBody struct {
-	Value         interface{} `json:"value"`
-	SchemaId      string      `json:"schemaId"`
-	SchemaVersion string      `json:"schemaVersion"`
-	Scope         string      `json:"scope,omitempty"`
+	Value         any    `json:"value"`
+	SchemaId      string `json:"schemaId"`
+	SchemaVersion string `json:"schemaVersion"`
+	Scope         string `json:"scope,omitempty"`
 }
 
 type postKubernetesAppSettings struct {
@@ -44,45 +44,8 @@ type kubernetesAppOptionsSettings struct {
 	EnableKubernetesApp bool `json:"enableKubernetesApp"`
 }
 
-type monitoredEntitiesResponse struct {
-	Entities   []MonitoredEntity `json:"entities"`
-	TotalCount int               `json:"totalCount"`
-	PageSize   int               `json:"pageSize"`
-}
-
-type MonitoredEntity struct {
-	EntityId    string `json:"entityId"`
-	DisplayName string `json:"displayName"`
-	LastSeenTms int64  `json:"lastSeenTms"`
-}
-
-type GetSettingsResponse struct {
-	TotalCount int `json:"totalCount"`
-}
-
-type postSettingsResponse struct {
-	ObjectId string `json:"objectId"`
-}
-
-type getSettingsErrorResponse struct {
-	ErrorMessage getSettingsError `json:"error"`
-}
-
-type getSettingsError struct {
-	Message              string
-	ConstraintViolations constraintViolations `json:"constraintViolations,omitempty"`
-	Code                 int
-}
-
-type constraintViolations []struct {
-	ParameterLocation string
-	Location          string
-	Message           string
-	Path              string
-}
-
 const (
-	SettingsSchemaId                            = "builtin:cloud.kubernetes"
+	KubernetesSettingsSchemaId                  = "builtin:cloud.kubernetes"
 	AppTransitionSchemaId                       = "builtin:app-transition.kubernetes"
 	schemaVersionV1                             = "1.0.27"
 	hierarchicalMonitoringSettingsSchemaVersion = "3.0.0"
@@ -165,7 +128,7 @@ func createV1KubernetesSettingsBody(clusterLabel, kubeSystemUUID, scope string) 
 	}
 	postK8sSettings.MonitoringSettings = &ms
 
-	settings := createBaseKubernetesSettings(postK8sSettings, SettingsSchemaId, schemaVersionV1, scope)
+	settings := createBaseKubernetesSettings(postK8sSettings, KubernetesSettingsSchemaId, schemaVersionV1, scope)
 
 	return []postKubernetesSettingsBody{settings}
 }
@@ -173,7 +136,7 @@ func createV1KubernetesSettingsBody(clusterLabel, kubeSystemUUID, scope string) 
 func createV3KubernetesSettingsBody(clusterLabel, kubeSystemUUID, scope string) []postKubernetesSettingsBody {
 	settings := createBaseKubernetesSettings(
 		createPostKubernetesSettings(clusterLabel, kubeSystemUUID),
-		SettingsSchemaId,
+		KubernetesSettingsSchemaId,
 		hierarchicalMonitoringSettingsSchemaVersion,
 		scope)
 	settings.SchemaVersion = hierarchicalMonitoringSettingsSchemaVersion
@@ -200,127 +163,6 @@ func (dtc *dynatraceClient) CreateOrUpdateKubernetesSetting(ctx context.Context,
 	}
 
 	return objectId, nil
-}
-
-func (dtc *dynatraceClient) GetMonitoredEntitiesForKubeSystemUUID(ctx context.Context, kubeSystemUUID string) ([]MonitoredEntity, error) {
-	if kubeSystemUUID == "" {
-		return nil, errors.New("no kube-system namespace UUID given")
-	}
-
-	req, err := createBaseRequest(ctx, dtc.getEntitiesUrl(), http.MethodGet, dtc.apiToken, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	q := req.URL.Query()
-	q.Add("pageSize", "500")
-	q.Add("entitySelector", fmt.Sprintf("type(KUBERNETES_CLUSTER),kubernetesClusterId(%s)", kubeSystemUUID))
-	q.Add("from", "-365d")
-	q.Add("fields", "+lastSeenTms")
-	req.URL.RawQuery = q.Encode()
-
-	res, err := dtc.httpClient.Do(req)
-	if err != nil {
-		log.Info("check if ME exists failed")
-
-		return nil, err
-	}
-
-	defer utils.CloseBodyAfterRequest(res)
-
-	var resDataJson monitoredEntitiesResponse
-
-	err = dtc.unmarshalToJson(res, &resDataJson)
-	if err != nil {
-		return nil, fmt.Errorf("error parsing response body: %w", err)
-	}
-
-	return resDataJson.Entities, nil
-}
-
-func (dtc *dynatraceClient) GetSettingsForMonitoredEntities(ctx context.Context, monitoredEntities []MonitoredEntity, schemaId string) (GetSettingsResponse, error) {
-	if len(monitoredEntities) < 1 {
-		return GetSettingsResponse{TotalCount: 0}, nil
-	}
-
-	scopes := make([]string, 0, len(monitoredEntities))
-	for _, entity := range monitoredEntities {
-		scopes = append(scopes, entity.EntityId)
-	}
-
-	req, err := createBaseRequest(ctx, dtc.getSettingsUrl(true), http.MethodGet, dtc.apiToken, nil)
-	if err != nil {
-		return GetSettingsResponse{}, err
-	}
-
-	q := req.URL.Query()
-	q.Add("schemaIds", schemaId)
-	q.Add("scopes", strings.Join(scopes, ","))
-	req.URL.RawQuery = q.Encode()
-
-	res, err := dtc.httpClient.Do(req)
-	if err != nil {
-		log.Info("failed to retrieve MEs")
-
-		return GetSettingsResponse{}, err
-	}
-
-	defer utils.CloseBodyAfterRequest(res)
-
-	var resDataJson GetSettingsResponse
-
-	err = dtc.unmarshalToJson(res, &resDataJson)
-	if err != nil {
-		return GetSettingsResponse{}, fmt.Errorf("error parsing response body: %w", err)
-	}
-
-	return resDataJson, nil
-}
-
-func (dtc *dynatraceClient) unmarshalToJson(res *http.Response, resDataJson interface{}) error {
-	resData, err := dtc.getServerResponseData(res)
-	if err != nil {
-		return fmt.Errorf("error reading response body: %w", err)
-	}
-
-	err = json.Unmarshal(resData, resDataJson)
-	if err != nil {
-		return fmt.Errorf("error parsing response body: %w", err)
-	}
-
-	return nil
-}
-
-func handleErrorArrayResponseFromAPI(response []byte, statusCode int) error {
-	if statusCode == http.StatusForbidden || statusCode == http.StatusUnauthorized {
-		var se getSettingsErrorResponse
-		if err := json.Unmarshal(response, &se); err != nil {
-			return fmt.Errorf("response error: %d, can't unmarshal json response", statusCode)
-		}
-
-		return fmt.Errorf("response error: %d, %s", statusCode, se.ErrorMessage.Message)
-	} else {
-		var se []getSettingsErrorResponse
-		if err := json.Unmarshal(response, &se); err != nil {
-			return fmt.Errorf("response error: %d, can't unmarshal json response", statusCode)
-		}
-
-		var sb strings.Builder
-
-		sb.WriteString("[Settings Creation]: could not create the Kubernetes setting for the following reason:\n")
-
-		for _, errorResponse := range se {
-			sb.WriteString(fmt.Sprintf("[%s; Code: %d\n", errorResponse.ErrorMessage.Message, errorResponse.ErrorMessage.Code))
-
-			for _, constraintViolation := range errorResponse.ErrorMessage.ConstraintViolations {
-				sb.WriteString(fmt.Sprintf("\t- %s\n", constraintViolation.Message))
-			}
-
-			sb.WriteString("]\n")
-		}
-
-		return errors.New(sb.String())
-	}
 }
 
 func (dtc *dynatraceClient) CreateOrUpdateKubernetesAppSetting(ctx context.Context, scope string) (string, error) {
