+++ pkg/clients/dynatrace/settings_kubernetes_test.go
@@ -3,7 +3,7 @@ package dynatrace
 import (
 	"context"
 	"encoding/json"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"strconv"
@@ -13,198 +13,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-const (
-	testUID      = "test-uid"
-	testName     = "test-name"
-	testObjectID = "test-objectid"
-	testScope    = "test-scope"
-)
-
-func TestDynatraceClient_GetMonitoredEntitiesForKubeSystemUUID(t *testing.T) {
-	ctx := context.Background()
-
-	t.Run("monitored entities for this uuid exist", func(t *testing.T) {
-		// arrange
-		expected := createMonitoredEntitiesForTesting()
-
-		dynatraceServer := httptest.NewServer(mockDynatraceServerEntitiesHandler(expected, false))
-		defer dynatraceServer.Close()
-
-		skipCert := SkipCertificateValidation(true)
-		dtc, err := NewClient(dynatraceServer.URL, apiToken, paasToken, skipCert)
-		require.NoError(t, err)
-		require.NotNil(t, dtc)
-
-		// act
-		actual, err := dtc.GetMonitoredEntitiesForKubeSystemUUID(ctx, testUID)
-
-		// assert
-		require.NotNil(t, actual)
-		require.NoError(t, err)
-		assert.Len(t, actual, 2)
-		assert.EqualValues(t, expected, actual)
-	})
-
-	t.Run("no monitored entities for this uuid exist", func(t *testing.T) {
-		// arrange
-		expected := []MonitoredEntity{}
-
-		dynatraceServer := httptest.NewServer(mockDynatraceServerEntitiesHandler(expected, false))
-		defer dynatraceServer.Close()
-
-		skipCert := SkipCertificateValidation(true)
-		dtc, err := NewClient(dynatraceServer.URL, apiToken, paasToken, skipCert)
-		require.NoError(t, err)
-		require.NotNil(t, dtc)
-
-		// act
-		actual, err := dtc.GetMonitoredEntitiesForKubeSystemUUID(ctx, testUID)
-
-		// assert
-		require.NotNil(t, actual)
-		require.NoError(t, err)
-		assert.Empty(t, actual)
-		assert.EqualValues(t, expected, actual)
-	})
-
-	t.Run("no monitored entities found because no kube-system uuid is provided", func(t *testing.T) {
-		// arrange
-		expected := createMonitoredEntitiesForTesting()
-
-		dynatraceServer := httptest.NewServer(mockDynatraceServerEntitiesHandler(expected, true))
-		defer dynatraceServer.Close()
-
-		skipCert := SkipCertificateValidation(true)
-		dtc, err := NewClient(dynatraceServer.URL, apiToken, paasToken, skipCert)
-		require.NoError(t, err)
-		require.NotNil(t, dtc)
-
-		// act
-		actual, err := dtc.GetMonitoredEntitiesForKubeSystemUUID(ctx, "")
-
-		// assert
-		require.Nil(t, actual)
-		require.Error(t, err)
-		assert.Empty(t, actual)
-	})
-
-	t.Run("no monitored entities found because of an api error", func(t *testing.T) {
-		// arrange
-		expected := createMonitoredEntitiesForTesting()
-
-		dynatraceServer := httptest.NewServer(mockDynatraceServerEntitiesHandler(expected, true))
-		defer dynatraceServer.Close()
-
-		skipCert := SkipCertificateValidation(true)
-		dtc, err := NewClient(dynatraceServer.URL, apiToken, paasToken, skipCert)
-		require.NoError(t, err)
-		require.NotNil(t, dtc)
-
-		// act
-		actual, err := dtc.GetMonitoredEntitiesForKubeSystemUUID(ctx, testUID)
-
-		// assert
-		require.Nil(t, actual)
-		require.Error(t, err)
-		assert.Empty(t, actual)
-	})
-}
-
-func TestDynatraceClient_GetSettingsForMonitoredEntities(t *testing.T) {
-	ctx := context.Background()
-
-	t.Run(`settings for the given monitored entities exist`, func(t *testing.T) {
-		// arrange
-		expected := createMonitoredEntitiesForTesting()
-		totalCount := 2
-
-		dynatraceServer := httptest.NewServer(mockDynatraceServerSettingsHandler(totalCount, "", http.StatusOK))
-		defer dynatraceServer.Close()
-
-		skipCert := SkipCertificateValidation(true)
-		dtc, err := NewClient(dynatraceServer.URL, apiToken, paasToken, skipCert)
-		require.NoError(t, err)
-		require.NotNil(t, dtc)
-
-		// act
-		actual, err := dtc.GetSettingsForMonitoredEntities(ctx, expected, SettingsSchemaId)
-
-		// assert
-		require.NoError(t, err)
-		require.NotNil(t, actual)
-		assert.Positive(t, actual.TotalCount)
-		assert.Len(t, expected, actual.TotalCount)
-	})
-
-	t.Run(`no settings for the given monitored entities exist`, func(t *testing.T) {
-		// arrange
-		expected := createMonitoredEntitiesForTesting()
-		totalCount := 0
-
-		dynatraceServer := httptest.NewServer(mockDynatraceServerSettingsHandler(totalCount, "", http.StatusOK))
-		defer dynatraceServer.Close()
-
-		skipCert := SkipCertificateValidation(true)
-		dtc, err := NewClient(dynatraceServer.URL, apiToken, paasToken, skipCert)
-		require.NoError(t, err)
-		require.NotNil(t, dtc)
-
-		// act
-		actual, err := dtc.GetSettingsForMonitoredEntities(ctx, expected, SettingsSchemaId)
-
-		// assert
-		require.NoError(t, err)
-		require.NotNil(t, actual)
-		assert.Less(t, actual.TotalCount, 1)
-	})
-
-	t.Run(`no settings for an empty list of monitored entities exist`, func(t *testing.T) {
-		// arrange
-		entities := []MonitoredEntity{}
-		// it is immaterial what we put here since no http call is executed when the list of
-		// monitored entities is empty, therefore also no settings will be returned
-		totalCount := 999
-
-		dynatraceServer := httptest.NewServer(mockDynatraceServerSettingsHandler(totalCount, "", http.StatusOK))
-		defer dynatraceServer.Close()
-
-		skipCert := SkipCertificateValidation(true)
-		dtc, err := NewClient(dynatraceServer.URL, apiToken, paasToken, skipCert)
-		require.NoError(t, err)
-		require.NotNil(t, dtc)
-
-		// act
-		actual, err := dtc.GetSettingsForMonitoredEntities(ctx, entities, SettingsSchemaId)
-
-		// assert
-		require.NoError(t, err)
-		require.NotNil(t, actual)
-		assert.Empty(t, actual.TotalCount)
-	})
-
-	t.Run(`no settings found for because of an api error`, func(t *testing.T) {
-		// arrange
-		entities := createMonitoredEntitiesForTesting()
-		// it is immaterial what we put here since the http request is producing an error
-		totalCount := 999
-
-		dynatraceServer := httptest.NewServer(mockDynatraceServerSettingsHandler(totalCount, "", http.StatusBadRequest))
-		defer dynatraceServer.Close()
-
-		skipCert := SkipCertificateValidation(true)
-		dtc, err := NewClient(dynatraceServer.URL, apiToken, paasToken, skipCert)
-		require.NoError(t, err)
-		require.NotNil(t, dtc)
-
-		// act
-		actual, err := dtc.GetSettingsForMonitoredEntities(ctx, entities, SettingsSchemaId)
-
-		// assert
-		require.Error(t, err)
-		assert.Empty(t, actual.TotalCount)
-	})
-}
 
 func TestDynatraceClient_CreateOrUpdateKubernetesSetting(t *testing.T) {
 	ctx := context.Background()
@@ -410,13 +218,6 @@ func TestDynatraceClient_getKubernetesSettingBody(t *testing.T) {
 	})
 }
 
-func createMonitoredEntitiesForTesting() []MonitoredEntity {
-	return []MonitoredEntity{
-		{EntityId: "KUBERNETES_CLUSTER-0E30FE4BF2007587", DisplayName: "operator test entity 1", LastSeenTms: 1639483869085},
-		{EntityId: "KUBERNETES_CLUSTER-119C75CCDA94799F", DisplayName: "operator test entity 2", LastSeenTms: 1639034988126},
-	}
-}
-
 func mockHandleEntitiesRequest(request *http.Request, writer http.ResponseWriter, entities []MonitoredEntity) {
 	if request.Method == http.MethodGet {
 		if !strings.Contains(request.Form.Get("entitySelector"), "type(KUBERNETES_CLUSTER)") {
@@ -466,7 +267,7 @@ func mockHandleSettingsRequest(request *http.Request, writer http.ResponseWriter
 			return
 		}
 
-		body, err := ioutil.ReadAll(request.Body)
+		body, err := io.ReadAll(request.Body)
 		if err != nil {
 			return
 		}
